+++ packages/lds/src/SearchInput/SearchInput.tsx
@@ -3,8 +3,9 @@
 import { typography } from '@linker/styles';
 import clsx from 'clsx';
 import Image from 'next/image';
-import { HTMLAttributes, useRef } from 'react';
+import { HTMLAttributes, useEffect, useRef } from 'react';
 import { Controller, useForm, useWatch } from 'react-hook-form';
+import { useDebounce } from 'use-debounce';
 
 import { searchInputContainer, searchInput } from './SearchInput.css';
 
@@ -12,23 +13,25 @@ interface Props extends HTMLAttributes<HTMLInputElement> {
   children?: React.ReactNode;
   className?: string;
   placeholder: string;
+  query: string;
+  setQuery: React.Dispatch<React.SetStateAction<string>>;
 }
 
-const SearchInput = ({ children, className, placeholder }: Props) => {
-  const { register, control, handleSubmit } = useForm();
-  const inputRef = useRef<HTMLInputElement>(null);
+const SearchInput = ({ children, className, placeholder, setQuery, query }: Props) => {
+  const { register, control } = useForm();
   const watchQuery = useWatch({
     name: 'search',
     control,
   });
+  const [value] = useDebounce(watchQuery, 1000);
+  const inputRef = useRef<HTMLInputElement>(null);
 
-  /*eslint-disable no-console*/
-  console.log(watchQuery);
-
-  const onSubmit = () => {};
+  useEffect(() => {
+    setQuery(value);
+  }, [watchQuery]);
 
   return (
-    <form className={searchInputContainer} onSubmit={handleSubmit(onSubmit)}>
+    <nav className={searchInputContainer}>
       <Image
         src="https://static.im-linker.com/search.svg"
         width={24}
@@ -50,7 +53,7 @@ const SearchInput = ({ children, className, placeholder }: Props) => {
           />
         )}
       />
-    </form>
+    </nav>
   );
 };
