+++ services/web/src/app/my/contact/Contact.css.ts
@@ -11,31 +11,3 @@ export const wrapper = style({
 export const searchInputWrapper = style({
   marginTop: '1.9rem',
 });
-
-export const favoritesWrapper = style({
-  paddingTop: '38px',
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'space-between',
-});
-export const favoritesTextWrapper = style({
-  display: 'flex',
-  flexDirection: 'row',
-  gap: '1px',
-});
-export const favoriteDropDown = style({
-  cursor: 'pointer',
-});
-export const totalCountWrapper = style({
-  paddingTop: '38px',
-  display: 'flex',
-  flexDirection: 'row',
-  gap: '1px',
-});
-
-export const profileWrapper = style({
-  display: 'flex',
-  flexDirection: 'column',
-  paddingTop: '26px',
-  gap: '16px',
-});
